MODPATBLK-86 Add PomUtils
